requirements.txt
aiocontextvars==0.2.2
aiofiles==0.4.0
aiosqlite==0.12.0
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==1.3.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
chart-studio==1.0.0
click==7.1.1
colorama==0.3.7
command-not-found==0.3
contextvars==2.4
cryptography==2.8
cycler==0.10.0
dashboard==0.0.2
databases==0.2.6
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0

django-crispy-forms==1.9.0
django-filter==2.2.0

entrypoints==0.3
filelock==3.0.12

h11==0.9.0
html5lib==0.999999999
httplib2==0.9.2
httptools==0.1.1
idna==2.8
immutables==0.11
importlib-metadata==1.3.0
importlib-resources==1.2.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.15.2
jeepney==0.4.2
Jinja2==2.10.3
joblib==0.14.1
json-tricks==3.15.2
jsonschema==3.2.0
keyring==21.1.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
louis==3.5.0
lxml==4.4.2
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.9
nose==1.3.7
notebook==6.0.2
numexpr==2.7.1
numpy==1.18.0
olefile==0.46
onboard==1.4.1
orm==0.2.0.dev0
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
PuLP==2.1
py==1.8.1
pycairo==1.16.2
pyclustertend==1.4.0
pycparser==2.19
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pyinotify==0.9.6
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
python-apt===1.6.5-ubuntu0.2-elementary14-ubuntu5.1.1
python-dateutil==2.8.1
python-debian==0.1.36
python-multipart==0.0.5
pytz==2019.3
pyxdg==0.26
PyYAML==3.12
pyzmq==18.1.1
qtconsole==4.6.0
reportlab==3.5.32
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
scikit-criteria==0.2.10
SecretStorage==3.1.2
Send2Trash==1.5.0
simplejson==3.17.0
six==1.13.0
SQLAlchemy==1.3.16
sqlparse==0.3.0
starlette==0.13.2
tables==3.6.1
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typesystem==0.3.0.dev0
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.7
usb-creator==0.3.3
uvicorn==0.11.3
uvloop==0.14.0
virtualenv==20.0.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
xkit==0.0.0
zipp==0.6.0
